Refactor: Move accordion components to their own directories
